requirements.txt
pyyaml
boto3
feedgen
flask
gunicorn
redis
rq
